Added ApolloConfig file for apollo requests.
