+++ app/src/main/res/layout/activity_details.xml
@@ -170,7 +170,6 @@
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/langRV"
             android:layout_width="wrap_content"
-
             android:layout_height="wrap_content" />
 
     </LinearLayout>
